checkin/requirements.txt
Flask
Flask_SQLAlchemy
Flask_Login
Flask_WTF
WTForms
Werkzeug
Flask_Bcrypt